Use stored offset as default for `load_kafka`
Changes the default offset from "end" to "stored" when loading from
Kafka. This aligns with typical consumer expectations where reading
resumes from the last committed position rather than skipping to
the end of the topic.

Also fixes the offset parameter type from int to int64_t for
consistency with the RdKafka API.

+++ plugins/kafka/include/kafka/configuration.hpp
@@ -85,7 +85,7 @@ public:
 private:
   class rebalancer : public RdKafka::RebalanceCb {
   public:
-    explicit rebalancer(int offset);
+    explicit rebalancer(int64_t offset);
 
     auto rebalance_cb(RdKafka::KafkaConsumer*, RdKafka::ErrorCode,
                       std::vector<RdKafka::TopicPartition*>&) -> void override;
